Extract tests for hasArity to own file
